functions are moved to header file
